harden runtime paths and split remaining oversized modules
