+++ src/subtitle_ocr/engine_core.rs
@@ -1,0 +1,480 @@
+struct SubtitleCandidate {
+    stream_index: i32,
+    language_tag: Option<String>,
+}
+
+#[derive(Debug, Clone)]
+struct SubtitleCue {
+    start_ms: i64,
+    end_ms: i64,
+    text: String,
+}
+
+#[derive(Debug, Clone)]
+struct OcrBoundingBox {
+    left: i32,
+    top: i32,
+    right: i32,
+    bottom: i32,
+}
+
+#[derive(Debug, Clone)]
+struct OcrLine {
+    text: String,
+    bbox: Option<OcrBoundingBox>,
+    score: Option<f32>,
+    color: Option<(u8, u8, u8)>,
+    italic: bool,
+}
+
+#[derive(Debug, Default)]
+struct OcrOutput {
+    lines: Vec<OcrLine>,
+}
+
+trait SubtitleConverter {
+    fn extract_lines(&mut self, image_path: &Path, language: &str) -> Result<OcrOutput>;
+}
+
+struct TesseractEngine;
+
+struct ExternalEngine {
+    command: String,
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+enum PpOcrVariant {
+    V3,
+    V4,
+}
+
+impl PpOcrVariant {
+    fn label(self) -> &'static str {
+        match self {
+            PpOcrVariant::V3 => "PP-OCRv3",
+            PpOcrVariant::V4 => "PP-OCRv4",
+        }
+    }
+
+    fn model_specs(self) -> (&'static ModelSpec, &'static ModelSpec, &'static ModelSpec) {
+        match self {
+            PpOcrVariant::V3 => (
+                &PPOCR_V3_DET_MODEL,
+                &PPOCR_V3_CLS_MODEL,
+                &PPOCR_V3_REC_MODEL,
+            ),
+            PpOcrVariant::V4 => (
+                &PPOCR_V4_DET_MODEL,
+                &PPOCR_V4_CLS_MODEL,
+                &PPOCR_V4_REC_MODEL,
+            ),
+        }
+    }
+}
+
+struct PpOcrEngine {
+    ocr: OcrLite,
+    variant: PpOcrVariant,
+}
+
+impl SubtitleConverter for TesseractEngine {
+    fn extract_lines(&mut self, image_path: &Path, language: &str) -> Result<OcrOutput> {
+        let text = run_tesseract(image_path, language)?;
+        if text.is_empty() {
+            return Ok(OcrOutput::default());
+        }
+        Ok(OcrOutput {
+            lines: vec![OcrLine {
+                text,
+                bbox: None,
+                score: None,
+                color: None,
+                italic: false,
+            }],
+        })
+    }
+}
+
+impl SubtitleConverter for ExternalEngine {
+    fn extract_lines(&mut self, image_path: &Path, language: &str) -> Result<OcrOutput> {
+        let text = run_external_ocr_command(image_path, language, &self.command)?;
+        if text.is_empty() {
+            return Ok(OcrOutput::default());
+        }
+        Ok(OcrOutput {
+            lines: vec![OcrLine {
+                text,
+                bbox: None,
+                score: None,
+                color: None,
+                italic: false,
+            }],
+        })
+    }
+}
+
+impl SubtitleConverter for PpOcrEngine {
+    fn extract_lines(&mut self, image_path: &Path, _language: &str) -> Result<OcrOutput> {
+        let img = load_image(image_path)?;
+        let result = self
+            .ocr
+            .detect(&img, 50, 1024, 0.5, 0.3, 1.6, false, false)
+            .map_err(|err| anyhow!("{} failed: {}", self.variant.label(), err))?;
+
+        let mut lines = Vec::with_capacity(result.text_blocks.len());
+        for block in result.text_blocks {
+            let text = normalize_utf8_text(&block.text);
+            if text.is_empty() {
+                continue;
+            }
+            let bbox = bounding_box_from_points(&block.box_points);
+            lines.push(OcrLine {
+                text,
+                bbox,
+                score: Some(block.text_score),
+                color: None,
+                italic: false,
+            });
+        }
+
+        let lines = merge_ocr_lines_with_spacing(lines);
+
+        Ok(OcrOutput { lines })
+    }
+}
+
+static ORT_ENV_INIT: OnceLock<()> = OnceLock::new();
+static ORT_ENV_GPU_AVAILABLE: OnceLock<bool> = OnceLock::new();
+static FORCE_CPU_EP: AtomicBool = AtomicBool::new(false);
+static TESSERACT_LANG_CACHE: OnceLock<Result<HashSet<String>>> = OnceLock::new();
+static LEGACY_NVIDIA_MAXWELL: OnceLock<bool> = OnceLock::new();
+static DISABLE_TESS_FALLBACK_LOGGED: AtomicBool = AtomicBool::new(false);
+
+pub fn convert_bitmap_subtitles(
+    input_file: &CStr,
+    work_dir: &Path,
+    sub_mode: SubMode,
+    default_language: Option<&str>,
+    ocr_engine: OcrEngine,
+    ocr_format: OcrFormat,
+    ocr_external_command: Option<&str>,
+) -> Result<Vec<OcrSubtitleTrack>> {
+    if matches!(sub_mode, SubMode::Skip) {
+        return Ok(Vec::new());
+    }
+
+    let candidates = discover_candidates(input_file, sub_mode)?;
+    if candidates.is_empty() {
+        return Ok(Vec::new());
+    }
+
+    if matches!(ocr_engine, OcrEngine::External) && ocr_external_command.is_none() {
+        bail!("--ocr-engine=external requires --ocr-external-command");
+    }
+
+    let input_path = input_file
+        .to_str()
+        .map_err(|_| anyhow!("Input path must be valid UTF-8 for OCR side pass"))?
+        .to_string();
+    let system_language = detect_system_ocr_language();
+    let video_dimensions = probe_video_dimensions(input_file);
+
+    let (resolved_engine, mut engine) = build_ocr_engine(ocr_engine, ocr_external_command)?;
+    let available_langs = if matches!(resolved_engine, OcrEngine::Tesseract) {
+        list_tesseract_languages().context(
+            "Failed to query Tesseract language packs. Install `tesseract-ocr` and required traineddata files.",
+        )?
+    } else {
+        HashSet::new()
+    };
+
+    let mut tracks = Vec::with_capacity(candidates.len());
+    for candidate in candidates {
+        let resolved_lang = resolve_ocr_language(
+            candidate.language_tag.as_deref(),
+            default_language,
+            system_language.as_deref(),
+            &available_langs,
+            resolved_engine,
+        );
+        let subtitle_path = work_dir.join(format!(
+            "stream-{}.{}",
+            candidate.stream_index,
+            ocr_format.extension()
+        ));
+        let cues = ocr_single_stream(
+            &input_path,
+            candidate.stream_index,
+            &resolved_lang,
+            work_dir,
+            ocr_format,
+            video_dimensions,
+            resolved_engine,
+            &mut *engine,
+        )?;
+
+        match ocr_format {
+            OcrFormat::Srt => write_srt(&subtitle_path, &cues)?,
+            OcrFormat::Ass => write_ass(&subtitle_path, &cues, video_dimensions)?,
+        }
+
+        info!(
+            "OCR subtitle stream {} -> '{}' ({} cues, format={:?})",
+            candidate.stream_index,
+            subtitle_path.display(),
+            cues.len(),
+            ocr_format
+        );
+
+        tracks.push(OcrSubtitleTrack {
+            language: resolved_lang,
+            subtitle_path,
+            format: ocr_format,
+        });
+    }
+
+    Ok(tracks)
+}
+
+impl OcrFormat {
+    fn extension(self) -> &'static str {
+        match self {
+            OcrFormat::Srt => "srt",
+            OcrFormat::Ass => "ass",
+        }
+    }
+}
+
+fn build_ocr_engine(
+    ocr_engine: OcrEngine,
+    ocr_external_command: Option<&str>,
+) -> Result<(OcrEngine, Box<dyn SubtitleConverter>)> {
+    match ocr_engine {
+        OcrEngine::Tesseract => Ok((OcrEngine::Tesseract, Box::new(TesseractEngine))),
+        OcrEngine::External => {
+            let command = ocr_external_command
+                .ok_or_else(|| anyhow!("missing OCR external command"))?
+                .to_string();
+            Ok((OcrEngine::External, Box::new(ExternalEngine { command })))
+        }
+        OcrEngine::PpOcrV3 => {
+            let variant = PpOcrVariant::V3;
+            let gpu_available = init_ort_environment()?;
+            if !gpu_available {
+                warn!(
+                    "{} running without GPU acceleration. \
+                     Install the CUDA/DirectML/CoreML runtime or set DPN_OCR_REQUIRE_GPU=1 to fail if GPU is required.",
+                    variant.label(),
+                );
+            }
+            let model_dir = resolve_model_dir()?;
+            let engine = init_ppocr_engine(&model_dir, require_gpu(), variant)?;
+            Ok((OcrEngine::PpOcrV3, Box::new(engine)))
+        }
+        OcrEngine::PpOcrV4 => {
+            let variant = PpOcrVariant::V4;
+            let gpu_available = init_ort_environment()?;
+            if !gpu_available {
+                warn!(
+                    "{} running without GPU acceleration. \
+                     Install the CUDA/DirectML/CoreML runtime or set DPN_OCR_REQUIRE_GPU=1 to fail if GPU is required.",
+                    variant.label(),
+                );
+            }
+            let model_dir = resolve_model_dir()?;
+            let engine = init_ppocr_engine(&model_dir, require_gpu(), variant)?;
+            Ok((OcrEngine::PpOcrV4, Box::new(engine)))
+        }
+        OcrEngine::Auto => {
+            let require_gpu = require_gpu();
+            let gpu_available = match init_ort_environment() {
+                Ok(available) => available,
+                Err(err) => {
+                    if require_gpu {
+                        return Err(err);
+                    }
+                    warn!("PP-OCR unavailable; falling back to Tesseract: {}", err);
+                    return Ok((OcrEngine::Tesseract, Box::new(TesseractEngine)));
+                }
+            };
+
+            let selected_engine = auto_engine_preference(gpu_available);
+            if matches!(selected_engine, OcrEngine::Tesseract) {
+                info!("Auto-selected Tesseract (no GPU execution provider available).");
+                return Ok((OcrEngine::Tesseract, Box::new(TesseractEngine)));
+            }
+            let variant = match selected_engine {
+                OcrEngine::PpOcrV3 => PpOcrVariant::V3,
+                OcrEngine::PpOcrV4 => PpOcrVariant::V4,
+                _ => PpOcrVariant::V4,
+            };
+
+            match (|| -> Result<PpOcrEngine> {
+                let model_dir = resolve_model_dir()?;
+                init_ppocr_engine(&model_dir, require_gpu, variant)
+            })() {
+                Ok(engine) => {
+                    info!(
+                        "Auto-selected {} engine with GPU acceleration",
+                        variant.label()
+                    );
+                    Ok((selected_engine, Box::new(engine)))
+                }
+                Err(err) => {
+                    if require_gpu {
+                        return Err(err);
+                    }
+                    warn!(
+                        "{} unavailable; falling back to Tesseract: {}",
+                        variant.label(),
+                        err
+                    );
+                    Ok((OcrEngine::Tesseract, Box::new(TesseractEngine)))
+                }
+            }
+        }
+    }
+}
+
+fn auto_engine_preference(gpu_available: bool) -> OcrEngine {
+    auto_engine_preference_with_capability(gpu_available, prefer_ppocr_v3_for_legacy_nvidia())
+}
+
+fn auto_engine_preference_with_capability(
+    gpu_available: bool,
+    prefer_v3_on_gpu: bool,
+) -> OcrEngine {
+    if !gpu_available {
+        return OcrEngine::Tesseract;
+    }
+    if prefer_v3_on_gpu {
+        OcrEngine::PpOcrV3
+    } else {
+        OcrEngine::PpOcrV4
+    }
+}
+
+fn prefer_ppocr_v3_for_legacy_nvidia() -> bool {
+    *LEGACY_NVIDIA_MAXWELL.get_or_init(detect_legacy_nvidia_maxwell)
+}
+
+fn detect_legacy_nvidia_maxwell() -> bool {
+    #[cfg(not(any(target_os = "linux", target_os = "windows")))]
+    {
+        return false;
+    }
+
+    #[cfg(any(target_os = "linux", target_os = "windows"))]
+    {
+        let output = Command::new("nvidia-smi")
+            .arg("--query-gpu=compute_cap,name")
+            .arg("--format=csv,noheader,nounits")
+            .output();
+
+        let Ok(output) = output else {
+            return false;
+        };
+        if !output.status.success() {
+            return false;
+        }
+
+        let stdout = String::from_utf8_lossy(&output.stdout);
+        for line in stdout
+            .lines()
+            .map(str::trim)
+            .filter(|line| !line.is_empty())
+        {
+            let mut parts = line.splitn(2, ',').map(str::trim);
+            let cap = parts.next().unwrap_or_default();
+            let name = parts.next().unwrap_or_default();
+            let major = cap
+                .split('.')
+                .next()
+                .and_then(|s| s.parse::<u32>().ok())
+                .unwrap_or(999);
+            if major <= 5 {
+                info!(
+                    "Detected legacy NVIDIA GPU '{}'(compute capability {}). Auto-selecting PP-OCRv3.",
+                    name,
+                    cap
+                );
+                return true;
+            }
+        }
+        false
+    }
+}
+
+fn disable_tesseract_quality_fallback() -> bool {
+    let disabled = env::var("DPN_OCR_DISABLE_TESS_FALLBACK")
+        .ok()
+        .map(|v| {
+            let x = v.trim().to_ascii_lowercase();
+            matches!(x.as_str(), "1" | "true" | "yes" | "on")
+        })
+        .unwrap_or(false);
+    if disabled && !DISABLE_TESS_FALLBACK_LOGGED.swap(true, Ordering::Relaxed) {
+        warn!(
+            "DPN_OCR_DISABLE_TESS_FALLBACK=1 set; skipping Tesseract quality fallback and using pure PP-OCR output."
+        );
+    }
+    disabled
+}
+
+fn ppocr_require_gpu_error(variant: PpOcrVariant, err: impl std::fmt::Display) -> anyhow::Error {
+    anyhow!(
+        "{} failed to initialize with DPN_OCR_REQUIRE_GPU=1. \
+         Verify CUDA/ONNX Runtime GPU libraries are installed. Underlying error: {}",
+        variant.label(),
+        err
+    )
+}
+
+fn init_ppocr_engine(
+    model_dir: &Path,
+    require_gpu: bool,
+    variant: PpOcrVariant,
+) -> Result<PpOcrEngine> {
+    let skip_cls = skip_ppocr_cls(variant, require_gpu);
+    if skip_cls {
+        info!(
+            "{} classifier model is disabled (DPN_OCR_SKIP_CLS or Maxwell GPU mode).",
+            variant.label()
+        );
+    }
+    match PpOcrEngine::new(model_dir, variant, skip_cls) {
+        Ok(engine) => Ok(engine),
+        Err(err) => {
+            if require_gpu {
+                return Err(ppocr_require_gpu_error(variant, err));
+            }
+            if force_cpu_execution_providers() {
+                return Err(err);
+            }
+            warn!(
+                "{} failed to initialize with GPU providers; retrying with CPU-only providers: {}",
+                variant.label(),
+                err
+            );
+            FORCE_CPU_EP.store(true, Ordering::Relaxed);
+            match PpOcrEngine::new(model_dir, variant, skip_cls) {
+                Ok(engine) => {
+                    info!(
+                        "{} initialized with CPU-only execution provider",
+                        variant.label()
+                    );
+                    Ok(engine)
+                }
+                Err(retry_err) => {
+                    warn!(
+                        "{} CPU-only initialization failed; falling back: {}",
+                        variant.label(),
+                        retry_err
+                    );
+                    Err(err)
+                }
+            }
+        }
+    }
+}
+
