+++ src/subtitle_ocr/engine_setup.rs
@@ -1,1098 +1,3 @@
-struct SubtitleCandidate {
-    stream_index: i32,
-    language_tag: Option<String>,
-}
-
-#[derive(Debug, Clone)]
-struct SubtitleCue {
-    start_ms: i64,
-    end_ms: i64,
-    text: String,
-}
-
-#[derive(Debug, Clone)]
-struct OcrBoundingBox {
-    left: i32,
-    top: i32,
-    right: i32,
-    bottom: i32,
-}
-
-#[derive(Debug, Clone)]
-struct OcrLine {
-    text: String,
-    bbox: Option<OcrBoundingBox>,
-    score: Option<f32>,
-    color: Option<(u8, u8, u8)>,
-    italic: bool,
-}
-
-#[derive(Debug, Default)]
-struct OcrOutput {
-    lines: Vec<OcrLine>,
-}
-
-trait SubtitleConverter {
-    fn extract_lines(&mut self, image_path: &Path, language: &str) -> Result<OcrOutput>;
-}
-
-struct TesseractEngine;
-
-struct ExternalEngine {
-    command: String,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum PpOcrVariant {
-    V3,
-    V4,
-}
-
-impl PpOcrVariant {
-    fn label(self) -> &'static str {
-        match self {
-            PpOcrVariant::V3 => "PP-OCRv3",
-            PpOcrVariant::V4 => "PP-OCRv4",
-        }
-    }
-
-    fn model_specs(self) -> (&'static ModelSpec, &'static ModelSpec, &'static ModelSpec) {
-        match self {
-            PpOcrVariant::V3 => (
-                &PPOCR_V3_DET_MODEL,
-                &PPOCR_V3_CLS_MODEL,
-                &PPOCR_V3_REC_MODEL,
-            ),
-            PpOcrVariant::V4 => (
-                &PPOCR_V4_DET_MODEL,
-                &PPOCR_V4_CLS_MODEL,
-                &PPOCR_V4_REC_MODEL,
-            ),
-        }
-    }
-}
-
-struct PpOcrEngine {
-    ocr: OcrLite,
-    variant: PpOcrVariant,
-}
-
-impl SubtitleConverter for TesseractEngine {
-    fn extract_lines(&mut self, image_path: &Path, language: &str) -> Result<OcrOutput> {
-        let text = run_tesseract(image_path, language)?;
-        if text.is_empty() {
-            return Ok(OcrOutput::default());
-        }
-        Ok(OcrOutput {
-            lines: vec![OcrLine {
-                text,
-                bbox: None,
-                score: None,
-                color: None,
-                italic: false,
-            }],
-        })
-    }
-}
-
-impl SubtitleConverter for ExternalEngine {
-    fn extract_lines(&mut self, image_path: &Path, language: &str) -> Result<OcrOutput> {
-        let text = run_external_ocr_command(image_path, language, &self.command)?;
-        if text.is_empty() {
-            return Ok(OcrOutput::default());
-        }
-        Ok(OcrOutput {
-            lines: vec![OcrLine {
-                text,
-                bbox: None,
-                score: None,
-                color: None,
-                italic: false,
-            }],
-        })
-    }
-}
-
-impl SubtitleConverter for PpOcrEngine {
-    fn extract_lines(&mut self, image_path: &Path, _language: &str) -> Result<OcrOutput> {
-        let img = load_image(image_path)?;
-        let result = self
-            .ocr
-            .detect(&img, 50, 1024, 0.5, 0.3, 1.6, false, false)
-            .map_err(|err| anyhow!("{} failed: {}", self.variant.label(), err))?;
-
-        let mut lines = Vec::with_capacity(result.text_blocks.len());
-        for block in result.text_blocks {
-            let text = normalize_utf8_text(&block.text);
-            if text.is_empty() {
-                continue;
-            }
-            let bbox = bounding_box_from_points(&block.box_points);
-            lines.push(OcrLine {
-                text,
-                bbox,
-                score: Some(block.text_score),
-                color: None,
-                italic: false,
-            });
-        }
-
-        let lines = merge_ocr_lines_with_spacing(lines);
-
-        Ok(OcrOutput { lines })
-    }
-}
-
-static ORT_ENV_INIT: OnceLock<()> = OnceLock::new();
-static ORT_ENV_GPU_AVAILABLE: OnceLock<bool> = OnceLock::new();
-static FORCE_CPU_EP: AtomicBool = AtomicBool::new(false);
-static TESSERACT_LANG_CACHE: OnceLock<Result<HashSet<String>>> = OnceLock::new();
-static LEGACY_NVIDIA_MAXWELL: OnceLock<bool> = OnceLock::new();
-static DISABLE_TESS_FALLBACK_LOGGED: AtomicBool = AtomicBool::new(false);
-
-pub fn convert_bitmap_subtitles(
-    input_file: &CStr,
-    work_dir: &Path,
-    sub_mode: SubMode,
-    default_language: Option<&str>,
-    ocr_engine: OcrEngine,
-    ocr_format: OcrFormat,
-    ocr_external_command: Option<&str>,
-) -> Result<Vec<OcrSubtitleTrack>> {
-    if matches!(sub_mode, SubMode::Skip) {
-        return Ok(Vec::new());
-    }
-
-    let candidates = discover_candidates(input_file, sub_mode)?;
-    if candidates.is_empty() {
-        return Ok(Vec::new());
-    }
-
-    if matches!(ocr_engine, OcrEngine::External) && ocr_external_command.is_none() {
-        bail!("--ocr-engine=external requires --ocr-external-command");
-    }
-
-    let input_path = input_file
-        .to_str()
-        .map_err(|_| anyhow!("Input path must be valid UTF-8 for OCR side pass"))?
-        .to_string();
-    let system_language = detect_system_ocr_language();
-    let video_dimensions = probe_video_dimensions(input_file);
-
-    let (resolved_engine, mut engine) = build_ocr_engine(ocr_engine, ocr_external_command)?;
-    let available_langs = if matches!(resolved_engine, OcrEngine::Tesseract) {
-        list_tesseract_languages().context(
-            "Failed to query Tesseract language packs. Install `tesseract-ocr` and required traineddata files.",
-        )?
-    } else {
-        HashSet::new()
-    };
-
-    let mut tracks = Vec::with_capacity(candidates.len());
-    for candidate in candidates {
-        let resolved_lang = resolve_ocr_language(
-            candidate.language_tag.as_deref(),
-            default_language,
-            system_language.as_deref(),
-            &available_langs,
-            resolved_engine,
-        );
-        let subtitle_path = work_dir.join(format!(
-            "stream-{}.{}",
-            candidate.stream_index,
-            ocr_format.extension()
-        ));
-        let cues = ocr_single_stream(
-            &input_path,
-            candidate.stream_index,
-            &resolved_lang,
-            work_dir,
-            ocr_format,
-            video_dimensions,
-            resolved_engine,
-            &mut *engine,
-        )?;
-
-        match ocr_format {
-            OcrFormat::Srt => write_srt(&subtitle_path, &cues)?,
-            OcrFormat::Ass => write_ass(&subtitle_path, &cues, video_dimensions)?,
-        }
-
-        info!(
-            "OCR subtitle stream {} -> '{}' ({} cues, format={:?})",
-            candidate.stream_index,
-            subtitle_path.display(),
-            cues.len(),
-            ocr_format
-        );
-
-        tracks.push(OcrSubtitleTrack {
-            language: resolved_lang,
-            subtitle_path,
-            format: ocr_format,
-        });
-    }
-
-    Ok(tracks)
-}
-
-impl OcrFormat {
-    fn extension(self) -> &'static str {
-        match self {
-            OcrFormat::Srt => "srt",
-            OcrFormat::Ass => "ass",
-        }
-    }
-}
-
-fn build_ocr_engine(
-    ocr_engine: OcrEngine,
-    ocr_external_command: Option<&str>,
-) -> Result<(OcrEngine, Box<dyn SubtitleConverter>)> {
-    match ocr_engine {
-        OcrEngine::Tesseract => Ok((OcrEngine::Tesseract, Box::new(TesseractEngine))),
-        OcrEngine::External => {
-            let command = ocr_external_command
-                .ok_or_else(|| anyhow!("missing OCR external command"))?
-                .to_string();
-            Ok((OcrEngine::External, Box::new(ExternalEngine { command })))
-        }
-        OcrEngine::PpOcrV3 => {
-            let variant = PpOcrVariant::V3;
-            let gpu_available = init_ort_environment()?;
-            if !gpu_available {
-                warn!(
-                    "{} running without GPU acceleration. \
-                     Install the CUDA/DirectML/CoreML runtime or set DPN_OCR_REQUIRE_GPU=1 to fail if GPU is required.",
-                    variant.label(),
-                );
-            }
-            let model_dir = resolve_model_dir()?;
-            let engine = init_ppocr_engine(&model_dir, require_gpu(), variant)?;
-            Ok((OcrEngine::PpOcrV3, Box::new(engine)))
-        }
-        OcrEngine::PpOcrV4 => {
-            let variant = PpOcrVariant::V4;
-            let gpu_available = init_ort_environment()?;
-            if !gpu_available {
-                warn!(
-                    "{} running without GPU acceleration. \
-                     Install the CUDA/DirectML/CoreML runtime or set DPN_OCR_REQUIRE_GPU=1 to fail if GPU is required.",
-                    variant.label(),
-                );
-            }
-            let model_dir = resolve_model_dir()?;
-            let engine = init_ppocr_engine(&model_dir, require_gpu(), variant)?;
-            Ok((OcrEngine::PpOcrV4, Box::new(engine)))
-        }
-        OcrEngine::Auto => {
-            let require_gpu = require_gpu();
-            let gpu_available = match init_ort_environment() {
-                Ok(available) => available,
-                Err(err) => {
-                    if require_gpu {
-                        return Err(err);
-                    }
-                    warn!("PP-OCR unavailable; falling back to Tesseract: {}", err);
-                    return Ok((OcrEngine::Tesseract, Box::new(TesseractEngine)));
-                }
-            };
-
-            let selected_engine = auto_engine_preference(gpu_available);
-            if matches!(selected_engine, OcrEngine::Tesseract) {
-                info!("Auto-selected Tesseract (no GPU execution provider available).");
-                return Ok((OcrEngine::Tesseract, Box::new(TesseractEngine)));
-            }
-            let variant = match selected_engine {
-                OcrEngine::PpOcrV3 => PpOcrVariant::V3,
-                OcrEngine::PpOcrV4 => PpOcrVariant::V4,
-                _ => PpOcrVariant::V4,
-            };
-
-            match (|| -> Result<PpOcrEngine> {
-                let model_dir = resolve_model_dir()?;
-                init_ppocr_engine(&model_dir, require_gpu, variant)
-            })() {
-                Ok(engine) => {
-                    info!(
-                        "Auto-selected {} engine with GPU acceleration",
-                        variant.label()
-                    );
-                    Ok((selected_engine, Box::new(engine)))
-                }
-                Err(err) => {
-                    if require_gpu {
-                        return Err(err);
-                    }
-                    warn!(
-                        "{} unavailable; falling back to Tesseract: {}",
-                        variant.label(),
-                        err
-                    );
-                    Ok((OcrEngine::Tesseract, Box::new(TesseractEngine)))
-                }
-            }
-        }
-    }
-}
-
-fn auto_engine_preference(gpu_available: bool) -> OcrEngine {
-    auto_engine_preference_with_capability(gpu_available, prefer_ppocr_v3_for_legacy_nvidia())
-}
-
-fn auto_engine_preference_with_capability(
-    gpu_available: bool,
-    prefer_v3_on_gpu: bool,
-) -> OcrEngine {
-    if !gpu_available {
-        return OcrEngine::Tesseract;
-    }
-    if prefer_v3_on_gpu {
-        OcrEngine::PpOcrV3
-    } else {
-        OcrEngine::PpOcrV4
-    }
-}
-
-fn prefer_ppocr_v3_for_legacy_nvidia() -> bool {
-    *LEGACY_NVIDIA_MAXWELL.get_or_init(detect_legacy_nvidia_maxwell)
-}
-
-fn detect_legacy_nvidia_maxwell() -> bool {
-    #[cfg(not(any(target_os = "linux", target_os = "windows")))]
-    {
-        return false;
-    }
-
-    #[cfg(any(target_os = "linux", target_os = "windows"))]
-    {
-        let output = Command::new("nvidia-smi")
-            .arg("--query-gpu=compute_cap,name")
-            .arg("--format=csv,noheader,nounits")
-            .output();
-
-        let Ok(output) = output else {
-            return false;
-        };
-        if !output.status.success() {
-            return false;
-        }
-
-        let stdout = String::from_utf8_lossy(&output.stdout);
-        for line in stdout
-            .lines()
-            .map(str::trim)
-            .filter(|line| !line.is_empty())
-        {
-            let mut parts = line.splitn(2, ',').map(str::trim);
-            let cap = parts.next().unwrap_or_default();
-            let name = parts.next().unwrap_or_default();
-            let major = cap
-                .split('.')
-                .next()
-                .and_then(|s| s.parse::<u32>().ok())
-                .unwrap_or(999);
-            if major <= 5 {
-                info!(
-                    "Detected legacy NVIDIA GPU '{}'(compute capability {}). Auto-selecting PP-OCRv3.",
-                    name,
-                    cap
-                );
-                return true;
-            }
-        }
-        false
-    }
-}
-
-fn disable_tesseract_quality_fallback() -> bool {
-    let disabled = env::var("DPN_OCR_DISABLE_TESS_FALLBACK")
-        .ok()
-        .map(|v| {
-            let x = v.trim().to_ascii_lowercase();
-            matches!(x.as_str(), "1" | "true" | "yes" | "on")
-        })
-        .unwrap_or(false);
-    if disabled && !DISABLE_TESS_FALLBACK_LOGGED.swap(true, Ordering::Relaxed) {
-        warn!(
-            "DPN_OCR_DISABLE_TESS_FALLBACK=1 set; skipping Tesseract quality fallback and using pure PP-OCR output."
-        );
-    }
-    disabled
-}
-
-fn ppocr_require_gpu_error(variant: PpOcrVariant, err: impl std::fmt::Display) -> anyhow::Error {
-    anyhow!(
-        "{} failed to initialize with DPN_OCR_REQUIRE_GPU=1. \
-         Verify CUDA/ONNX Runtime GPU libraries are installed. Underlying error: {}",
-        variant.label(),
-        err
-    )
-}
-
-fn init_ppocr_engine(
-    model_dir: &Path,
-    require_gpu: bool,
-    variant: PpOcrVariant,
-) -> Result<PpOcrEngine> {
-    let skip_cls = skip_ppocr_cls(variant, require_gpu);
-    if skip_cls {
-        info!(
-            "{} classifier model is disabled (DPN_OCR_SKIP_CLS or Maxwell GPU mode).",
-            variant.label()
-        );
-    }
-    match PpOcrEngine::new(model_dir, variant, skip_cls) {
-        Ok(engine) => Ok(engine),
-        Err(err) => {
-            if require_gpu {
-                return Err(ppocr_require_gpu_error(variant, err));
-            }
-            if force_cpu_execution_providers() {
-                return Err(err);
-            }
-            warn!(
-                "{} failed to initialize with GPU providers; retrying with CPU-only providers: {}",
-                variant.label(),
-                err
-            );
-            FORCE_CPU_EP.store(true, Ordering::Relaxed);
-            match PpOcrEngine::new(model_dir, variant, skip_cls) {
-                Ok(engine) => {
-                    info!(
-                        "{} initialized with CPU-only execution provider",
-                        variant.label()
-                    );
-                    Ok(engine)
-                }
-                Err(retry_err) => {
-                    warn!(
-                        "{} CPU-only initialization failed; falling back: {}",
-                        variant.label(),
-                        retry_err
-                    );
-                    Err(err)
-                }
-            }
-        }
-    }
-}
-
-fn init_ort_environment() -> Result<bool> {
-    if ORT_ENV_INIT.get().is_some() {
-        return Ok(*ORT_ENV_GPU_AVAILABLE.get().unwrap_or(&false));
-    }
-    let selection = build_execution_providers()?;
-    match ort::init().commit() {
-        Ok(true) => info!("Initialized ONNX Runtime environment for OCR execution providers"),
-        Ok(false) => debug!("ONNX Runtime environment already initialized; skipping reconfigure"),
-        Err(err) => {
-            warn!("Failed to initialize ONNX Runtime environment: {}", err);
-            return Err(anyhow!(
-                "Failed to initialize ONNX Runtime environment: {err}"
-            ));
-        }
-    }
-    let _ = ORT_ENV_INIT.set(());
-    let _ = ORT_ENV_GPU_AVAILABLE.set(selection.gpu_available);
-    Ok(selection.gpu_available)
-}
-
-struct ModelSpec {
-    filename: &'static str,
-    url: &'static str,
-    sha256: &'static str,
-}
-
-const PPOCR_V4_DET_MODEL: ModelSpec = ModelSpec {
-    filename: "ch_PP-OCRv4_det_infer.onnx",
-    url: "https://www.modelscope.cn/models/RapidAI/RapidOCR/resolve/v3.7.0/onnx/PP-OCRv4/det/ch_PP-OCRv4_det_infer.onnx",
-    sha256: "D2A7720D45A54257208B1E13E36A8479894CB74155A5EFE29462512D42F49DA9",
-};
-const PPOCR_V4_CLS_MODEL: ModelSpec = ModelSpec {
-    filename: "ch_ppocr_mobile_v2.0_cls_infer.onnx",
-    url: "https://www.modelscope.cn/models/RapidAI/RapidOCR/resolve/v3.7.0/onnx/PP-OCRv4/cls/ch_ppocr_mobile_v2.0_cls_infer.onnx",
-    sha256: "E47ACEDF663230F8863FF1AB0E64DD2D82B838FCEB5957146DAB185A89D6215C",
-};
-const PPOCR_V4_REC_MODEL: ModelSpec = ModelSpec {
-    filename: "en_PP-OCRv4_rec_infer.onnx",
-    url: "https://www.modelscope.cn/models/RapidAI/RapidOCR/resolve/v3.7.0/onnx/PP-OCRv4/rec/en_PP-OCRv4_rec_infer.onnx",
-    sha256: "E8770C967605983D1570CDF5352041DFB68FA0C21664F49F47B155ABD3E0E318",
-};
-
-const PPOCR_V3_DET_MODEL: ModelSpec = ModelSpec {
-    filename: "ch_PP-OCRv3_det_infer.onnx",
-    url: "https://huggingface.co/SWHL/RapidOCR/resolve/main/PP-OCRv3/ch_PP-OCRv3_det_infer.onnx?download=true",
-    sha256: "3439588C030FAEA393A54515F51E983D8E155B19A2E8ABA7891934C1CF0DE526",
-};
-const PPOCR_V3_CLS_MODEL: ModelSpec = ModelSpec {
-    filename: "ch_ppocr_mobile_v2.0_cls_train.onnx",
-    url: "https://huggingface.co/SWHL/RapidOCR/resolve/main/PP-OCRv3/ch_ppocr_mobile_v2.0_cls_train.onnx?download=true",
-    sha256: "70581B300B83BABD9E0DD1D7D74C5B006869E8796DA277A70C2E405BF9D77C82",
-};
-const PPOCR_V3_REC_MODEL: ModelSpec = ModelSpec {
-    filename: "en_PP-OCRv3_rec_infer.onnx",
-    url: "https://huggingface.co/SWHL/RapidOCR/resolve/main/PP-OCRv3/en_PP-OCRv3_rec_infer.onnx?download=true",
-    sha256: "EF7ABD8BD3629AE57EA2C28B425C1BD258A871B93FD2FE7C433946ADE9B5D9EA",
-};
-
-struct PpOcrModels {
-    det: PathBuf,
-    cls: PathBuf,
-    rec: PathBuf,
-}
-
-impl PpOcrEngine {
-    fn new(model_dir: &Path, variant: PpOcrVariant, skip_cls: bool) -> Result<Self> {
-        let models = ensure_ppocr_models(model_dir, variant, skip_cls)?;
-        info!(
-            "Initializing {} models (det='{}', cls='{}', rec='{}')",
-            variant.label(),
-            models.det.display(),
-            models.cls.display(),
-            models.rec.display()
-        );
-        let mut ocr = OcrLite::new();
-        ocr.init_models_custom(
-            models.det.to_string_lossy().as_ref(),
-            models.cls.to_string_lossy().as_ref(),
-            models.rec.to_string_lossy().as_ref(),
-            configure_ort_builder,
-        )
-        .map_err(|err| anyhow!("failed to initialize {} models: {}", variant.label(), err))?;
-        Ok(Self { ocr, variant })
-    }
-}
-
-fn configure_ort_builder(builder: SessionBuilder) -> Result<SessionBuilder, ort::Error> {
-    let selection = build_execution_providers().map_err(|err| ort::Error::new(err.to_string()))?;
-    let mut builder = builder.with_execution_providers(selection.providers)?;
-    builder = builder.with_intra_threads(
-        std::thread::available_parallelism()
-            .map(|n| n.get())
-            .unwrap_or(2),
-    )?;
-    builder.with_inter_threads(
-        std::thread::available_parallelism()
-            .map(|n| n.get())
-            .unwrap_or(2),
-    )
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum ExecutionProviderKind {
-    Cuda,
-    DirectML,
-    CoreML,
-    Cpu,
-}
-
-impl ExecutionProviderKind {
-    fn label(self) -> &'static str {
-        match self {
-            ExecutionProviderKind::Cuda => "cuda",
-            ExecutionProviderKind::DirectML => "directml",
-            ExecutionProviderKind::CoreML => "coreml",
-            ExecutionProviderKind::Cpu => "cpu",
-        }
-    }
-}
-
-struct ExecutionProviderSelection {
-    providers: Vec<ExecutionProviderDispatch>,
-    gpu_available: bool,
-}
-
-fn require_gpu() -> bool {
-    env::var("DPN_OCR_REQUIRE_GPU").ok().as_deref() == Some("1")
-}
-
-fn skip_ppocr_cls(variant: PpOcrVariant, require_gpu: bool) -> bool {
-    let configured = env::var("DPN_OCR_SKIP_CLS").ok().and_then(|v| {
-        let v = v.trim();
-        if v.eq_ignore_ascii_case("1")
-            || v.eq_ignore_ascii_case("true")
-            || v.eq_ignore_ascii_case("yes")
-            || v.eq_ignore_ascii_case("on")
-        {
-            Some(true)
-        } else if v.eq_ignore_ascii_case("0")
-            || v.eq_ignore_ascii_case("false")
-            || v.eq_ignore_ascii_case("no")
-            || v.eq_ignore_ascii_case("off")
-        {
-            Some(false)
-        } else {
-            None
-        }
-    });
-    configured.unwrap_or(matches!(variant, PpOcrVariant::V3) && require_gpu)
-}
-
-fn force_cpu_execution_providers() -> bool {
-    if env::var("DPN_OCR_FORCE_CPU").ok().as_deref() == Some("1") {
-        return true;
-    }
-    FORCE_CPU_EP.load(Ordering::Relaxed)
-}
-
-fn format_provider_kinds(kinds: &[ExecutionProviderKind]) -> String {
-    kinds
-        .iter()
-        .map(|kind| kind.label())
-        .collect::<Vec<_>>()
-        .join(", ")
-}
-
-#[cfg(any(target_os = "linux", target_os = "windows"))]
-fn apply_cuda_env_overrides(mut ep: CUDAExecutionProvider) -> CUDAExecutionProvider {
-    let flags = match env::var("ORT_CUDA_FLAGS") {
-        Ok(val) => val,
-        Err(_) => return ep,
-    };
-
-    for chunk in flags.split(',') {
-        let chunk = chunk.trim();
-        if chunk.is_empty() {
-            continue;
-        }
-        let Some((key, value)) = chunk.split_once('=') else {
-            continue;
-        };
-        let key = key.trim();
-        let value = value.trim();
-        match key {
-            "device_id" => {
-                if let Ok(parsed) = value.parse::<i32>() {
-                    ep = ep.with_device_id(parsed);
-                    info!("Applied ORT_CUDA_FLAGS device_id={}", parsed);
-                } else {
-                    warn!("Invalid ORT_CUDA_FLAGS device_id '{}'; ignoring.", value);
-                }
-            }
-            "gpu_mem_limit" => {
-                if let Ok(parsed) = value.parse::<usize>() {
-                    ep = ep.with_memory_limit(parsed);
-                    info!("Applied ORT_CUDA_FLAGS gpu_mem_limit={}", parsed);
-                } else {
-                    warn!(
-                        "Invalid ORT_CUDA_FLAGS gpu_mem_limit '{}'; ignoring.",
-                        value
-                    );
-                }
-            }
-            _ => {}
-        }
-    }
-
-    ep
-}
-
-#[cfg(any(target_os = "linux", target_os = "windows"))]
-fn build_cuda_provider(require_gpu: bool) -> ExecutionProviderDispatch {
-    let mut ep = CUDAExecutionProvider::default();
-    ep = apply_cuda_env_overrides(ep);
-    ep = ep
-        .with_conv_algorithm_search(CuDNNConvAlgorithmSearch::Heuristic)
-        .with_cuda_graph(false)
-        .with_conv_max_workspace(false)
-        .with_arena_extend_strategy(ArenaExtendStrategy::SameAsRequested);
-    info!("CUDA EP safety brakes enabled for Maxwell-class GPUs.");
-    let mut ep = ep.build();
-    if require_gpu {
-        ep = ep.error_on_failure();
-    }
-    ep
-}
-
-#[cfg(not(any(target_os = "linux", target_os = "windows")))]
-fn build_cuda_provider(_require_gpu: bool) -> ExecutionProviderDispatch {
-    unreachable!("CUDA execution provider is not supported on this platform");
-}
-
-#[cfg(target_os = "windows")]
-fn build_directml_provider(require_gpu: bool) -> ExecutionProviderDispatch {
-    let mut ep = DirectMLExecutionProvider::default().build();
-    if require_gpu {
-        ep = ep.error_on_failure();
-    }
-    ep
-}
-
-#[cfg(not(target_os = "windows"))]
-fn build_directml_provider(_require_gpu: bool) -> ExecutionProviderDispatch {
-    unreachable!("DirectML execution provider is only supported on Windows");
-}
-
-#[cfg(target_vendor = "apple")]
-fn build_coreml_provider(require_gpu: bool) -> ExecutionProviderDispatch {
-    let mut ep = CoreMLExecutionProvider::default().build();
-    if require_gpu {
-        ep = ep.error_on_failure();
-    }
-    ep
-}
-
-#[cfg(not(target_vendor = "apple"))]
-fn build_coreml_provider(_require_gpu: bool) -> ExecutionProviderDispatch {
-    unreachable!("CoreML execution provider is only supported on Apple platforms");
-}
-
-#[cfg(target_os = "windows")]
-fn detect_directml_available(force_cpu: bool) -> bool {
-    if force_cpu {
-        return false;
-    }
-    let directml = DirectMLExecutionProvider::default();
-    match directml.is_available() {
-        Ok(true) => {
-            info!("ONNX Runtime DirectML execution provider available");
-            true
-        }
-        Ok(false) => {
-            warn!("ONNX Runtime DirectML execution provider not available; falling back to CPU.");
-            false
-        }
-        Err(err) => {
-            warn!(
-                "Failed to query DirectML execution provider availability: {}",
-                err
-            );
-            false
-        }
-    }
-}
-
-#[cfg(not(target_os = "windows"))]
-fn detect_directml_available(_force_cpu: bool) -> bool {
-    false
-}
-
-#[cfg(target_vendor = "apple")]
-fn detect_coreml_available(force_cpu: bool) -> bool {
-    if force_cpu {
-        return false;
-    }
-    let coreml = CoreMLExecutionProvider::default();
-    match coreml.is_available() {
-        Ok(true) => {
-            info!("ONNX Runtime CoreML execution provider available");
-            true
-        }
-        Ok(false) => {
-            warn!("ONNX Runtime CoreML execution provider not available; falling back to CPU.");
-            false
-        }
-        Err(err) => {
-            warn!(
-                "Failed to query CoreML execution provider availability: {}",
-                err
-            );
-            false
-        }
-    }
-}
-
-#[cfg(not(target_vendor = "apple"))]
-fn detect_coreml_available(_force_cpu: bool) -> bool {
-    false
-}
-
-fn select_execution_provider_plan(
-    require_gpu: bool,
-    cuda_available: bool,
-    directml_available: bool,
-    coreml_available: bool,
-) -> Result<(Vec<ExecutionProviderKind>, bool)> {
-    let mut kinds = Vec::new();
-    let mut gpu_available = false;
-
-    if cuda_available {
-        gpu_available = true;
-        kinds.push(ExecutionProviderKind::Cuda);
-    }
-    if directml_available {
-        gpu_available = true;
-        kinds.push(ExecutionProviderKind::DirectML);
-    }
-    if coreml_available {
-        gpu_available = true;
-        kinds.push(ExecutionProviderKind::CoreML);
-    }
-
-    if require_gpu && !gpu_available {
-        bail!("No GPU execution providers available. Install the required GPU runtime libraries or unset DPN_OCR_REQUIRE_GPU=1 to allow CPU fallback.");
-    }
-
-    kinds.push(ExecutionProviderKind::Cpu);
-    Ok((kinds, gpu_available))
-}
-
-fn build_execution_providers() -> Result<ExecutionProviderSelection> {
-    let require_gpu = require_gpu();
-    if require_gpu {
-        info!("DPN_OCR_REQUIRE_GPU=1; GPU execution provider is required.");
-    }
-    let force_cpu = force_cpu_execution_providers();
-    if force_cpu {
-        warn!("DPN_OCR_FORCE_CPU=1; disabling GPU execution providers.");
-    }
-    if require_gpu && force_cpu {
-        bail!("DPN_OCR_REQUIRE_GPU=1 cannot be used with DPN_OCR_FORCE_CPU=1.");
-    }
-    let mut providers = Vec::new();
-    let cuda_available = if force_cpu {
-        false
-    } else {
-        #[cfg(any(target_os = "linux", target_os = "windows"))]
-        {
-            let cuda = CUDAExecutionProvider::default();
-            match cuda.is_available() {
-                Ok(true) => {
-                    info!("ONNX Runtime CUDA execution provider available");
-                    true
-                }
-                Ok(false) => {
-                    warn!(
-                        "ONNX Runtime CUDA execution provider not available; falling back to CPU."
-                    );
-                    false
-                }
-                Err(err) => {
-                    warn!(
-                        "Failed to query CUDA execution provider availability: {}",
-                        err
-                    );
-                    false
-                }
-            }
-        }
-        #[cfg(not(any(target_os = "linux", target_os = "windows")))]
-        {
-            false
-        }
-    };
-
-    let directml_available = detect_directml_available(force_cpu);
-    let coreml_available = detect_coreml_available(force_cpu);
-
-    let (kinds, gpu_available) = select_execution_provider_plan(
-        require_gpu,
-        cuda_available,
-        directml_available,
-        coreml_available,
-    )?;
-
-    if gpu_available {
-        info!(
-            "OCR execution provider order: {}",
-            format_provider_kinds(&kinds)
-        );
-    } else if force_cpu {
-        info!("OCR execution provider order: cpu (GPU providers disabled)");
-    } else {
-        warn!(
-            "No GPU execution providers available; PP-OCR will run on CPU. \
-             Install the GPU runtime libraries or set DPN_OCR_REQUIRE_GPU=1 to fail instead."
-        );
-    }
-
-    for kind in kinds {
-        match kind {
-            ExecutionProviderKind::Cuda => {
-                providers.push(build_cuda_provider(require_gpu));
-            }
-            ExecutionProviderKind::DirectML => {
-                providers.push(build_directml_provider(require_gpu));
-            }
-            ExecutionProviderKind::CoreML => {
-                providers.push(build_coreml_provider(require_gpu));
-            }
-            ExecutionProviderKind::Cpu => {
-                providers.push(CPUExecutionProvider::default().build());
-            }
-        }
-    }
-
-    Ok(ExecutionProviderSelection {
-        providers,
-        gpu_available,
-    })
-}
-
-fn resolve_model_dir() -> Result<PathBuf> {
-    if let Some(dir) = env::var_os("DPN_OCR_MODEL_DIR") {
-        let path = PathBuf::from(dir);
-        fs::create_dir_all(&path)
-            .with_context(|| format!("creating OCR model directory '{}'", path.display()))?;
-        return Ok(path);
-    }
-
-    if let Ok(exe) = env::current_exe() {
-        if let Some(parent) = exe.parent() {
-            let candidate = parent.join("models");
-            if candidate.is_dir() {
-                return Ok(candidate);
-            }
-        }
-    }
-
-    let fallback = if let Some(xdg) = env::var_os("XDG_CONFIG_HOME") {
-        PathBuf::from(xdg).join("direct-play-nice").join("models")
-    } else if let Some(home) = env::var_os("HOME") {
-        PathBuf::from(home)
-            .join(".config")
-            .join("direct-play-nice")
-            .join("models")
-    } else {
-        env::temp_dir().join("direct-play-nice-models")
-    };
-
-    fs::create_dir_all(&fallback)
-        .with_context(|| format!("creating OCR model directory '{}'", fallback.display()))?;
-    Ok(fallback)
-}
-
-fn ensure_ppocr_models(
-    model_dir: &Path,
-    variant: PpOcrVariant,
-    skip_cls: bool,
-) -> Result<PpOcrModels> {
-    let (det_spec, cls_spec, rec_spec) = variant.model_specs();
-    if skip_cls {
-        debug!(
-            "Skipping classifier is requested, but this build uses mandatory classifier initialization; loading cls model."
-        );
-    }
-    let det = ensure_model_file(model_dir, det_spec)?;
-    let cls = ensure_model_file(model_dir, cls_spec)?;
-    let rec = ensure_model_file(model_dir, rec_spec)?;
-    Ok(PpOcrModels { det, cls, rec })
-}
-
-fn ensure_model_file(model_dir: &Path, spec: &ModelSpec) -> Result<PathBuf> {
-    let path = model_dir.join(spec.filename);
-    if path.exists() {
-        if let Ok(hash) = sha256_file(&path) {
-            if hash != spec.sha256.to_ascii_lowercase() {
-                warn!(
-                    "OCR model '{}' hash mismatch (got {}, expected {}). Using existing file anyway.",
-                    path.display(),
-                    hash,
-                    spec.sha256
-                );
-            }
-        }
-        return Ok(path);
-    }
-
-    info!(
-        "Downloading OCR model '{}' from {}",
-        spec.filename, spec.url
-    );
-    download_model_with_values(&path, spec.url, spec.sha256, spec.filename)?;
-    Ok(path)
-}
-
-#[cfg(test)]
-fn ensure_model_file_with_values(
-    model_dir: &Path,
-    filename: &str,
-    url: &str,
-    sha256: &str,
-) -> Result<PathBuf> {
-    let path = model_dir.join(filename);
-    if path.exists() {
-        return Ok(path);
-    }
-    info!("Downloading OCR model '{}' from {}", filename, url);
-    download_model_with_values(&path, url, sha256, filename)?;
-    Ok(path)
-}
-
-fn download_model_with_values(path: &Path, url: &str, sha256: &str, filename: &str) -> Result<()> {
-    let parent = path
-        .parent()
-        .ok_or_else(|| anyhow!("invalid OCR model path '{}'", path.display()))?;
-    fs::create_dir_all(parent)
-        .with_context(|| format!("creating OCR model directory '{}'", parent.display()))?;
-
-    let tmp_path = path.with_extension("download");
-    let response = ureq::get(url)
-        .call()
-        .with_context(|| format!("downloading OCR model from {}", url))?;
-    let mut reader = response.into_reader();
-    let mut file = fs::File::create(&tmp_path)
-        .with_context(|| format!("creating '{}'", tmp_path.display()))?;
-    let mut hasher = Sha256::new();
-    let mut buf = [0u8; 16 * 1024];
-    let mut download_result = Ok(());
-    loop {
-        let read = match reader.read(&mut buf) {
-            Ok(read) => read,
-            Err(err) => {
-                download_result = Err(anyhow!(err).context("reading OCR model download"));
-                break;
-            }
-        };
-        if read == 0 {
-            break;
-        }
-        hasher.update(&buf[..read]);
-        if let Err(err) = file.write_all(&buf[..read]) {
-            download_result = Err(anyhow!(err).context("writing OCR model download"));
-            break;
-        }
-    }
-    let _ = file.flush();
-
-    if let Err(err) = download_result {
-        let _ = fs::remove_file(&tmp_path);
-        return Err(err);
-    }
-
-    let actual_hash = to_hex_lower(&hasher.finalize());
-    let expected = sha256.to_ascii_lowercase();
-    if actual_hash != expected {
-        let _ = fs::remove_file(&tmp_path);
-        bail!(
-            "downloaded OCR model '{}' hash mismatch (got {}, expected {})",
-            filename,
-            actual_hash,
-            expected
-        );
-    }
-
-    fs::rename(&tmp_path, path).with_context(|| {
-        format!(
-            "renaming OCR model download '{}' -> '{}'",
-            tmp_path.display(),
-            path.display()
-        )
-    })?;
-    Ok(())
-}
-
-fn sha256_file(path: &Path) -> Result<String> {
-    let mut file = fs::File::open(path)?;
-    let mut hasher = Sha256::new();
-    let mut buf = [0u8; 16 * 1024];
-    loop {
-        let read = file.read(&mut buf)?;
-        if read == 0 {
-            break;
-        }
-        hasher.update(&buf[..read]);
-    }
-    Ok(to_hex_lower(&hasher.finalize()))
-}
-
-fn to_hex_lower(bytes: &[u8]) -> String {
-    let mut out = String::with_capacity(bytes.len() * 2);
-    for byte in bytes {
-        use std::fmt::Write as _;
-        let _ = write!(&mut out, "{:02x}", byte);
-    }
-    out
-}
-
+include!("engine_core.rs");
+include!("engine_runtime.rs");
+include!("engine_models.rs");
